Added application level scoped session for testing
We have had problems when testing because the entity returned
from a session no longer works after a request is sent through
the WSGI application.  This is because Session.remove() is called
at the end of a response cycle.  Therefore, a testing sequence that
looks like:

    u = User.get(123)
    r = testapp.get('/user/%s' % u.id)
    assert u.username in r

will fail with an exception like:

    DetachedInstanceError: Instance <User at 0x...> is not bound to a Session;
    attribute refresh operation cannot proceed

The user instance (u) is tied to a session that gets removed as part of the GET
request, so the attributes of u are no longer available.

This commit changes this by creating an application level session and uses
that session when in a non-request context.  Therefore, in an example like
the above, the user instance (u) is attached to a different session than the
one that is being accessed by the request/response code.  Therefore, the user
entity stays accessible even after the response cycle is completed.